+++ src/artm/regularizer/hierarchy_sparsing_theta.h
@@ -60,7 +60,7 @@ class HierarchySparsingThetaAgent : public RegularizeThetaAgent {
   std::vector<float> topic_weight;              // tau for every topic
   std::vector<float> alpha_weight;              // iteration coef
   std::vector<float> parent_topic_proportion;   // p(supertopic)
-  double prior_parent_topic_probability;        // 1.0 / count(supertopics)
+  float prior_parent_topic_probability;         // 1.0 / count(supertopics)
   bool regularization_on = true;                // true if current batch is parent phi batch
 };
 
@@ -69,7 +69,7 @@ class HierarchySparsingTheta : public RegularizerInterface {
   explicit HierarchySparsingTheta(const HierarchySparsingThetaConfig& config);
 
   virtual std::shared_ptr<RegularizeThetaAgent>
-    CreateRegularizeThetaAgent(const Batch& batch, const ProcessBatchesArgs& args, double tau);
+    CreateRegularizeThetaAgent(const Batch& batch, const ProcessBatchesArgs& args, float tau);
 
   virtual google::protobuf::RepeatedPtrField<std::string> topics_to_regularize();
