+++ src/artm/regularizer/smooth_ptdw.h
@@ -21,10 +21,10 @@ class SmoothPtdwAgent : public RegularizePtdwAgent {
   friend class SmoothPtdw;
   SmoothPtdwConfig config_;
   ProcessBatchesArgs args_;
-  double tau_;
+  float tau_;
 
  public:
-  SmoothPtdwAgent(const SmoothPtdwConfig& config, const ProcessBatchesArgs& args, double tau)
+  SmoothPtdwAgent(const SmoothPtdwConfig& config, const ProcessBatchesArgs& args, float tau)
       : config_(config), args_(args), tau_(tau) {}
 
   virtual void Apply(int item_index, int inner_iter, ::artm::utility::LocalPhiMatrix<float>* ptdw) const;
@@ -36,7 +36,7 @@ class SmoothPtdw : public RegularizerInterface {
     : config_(config) {}
 
   virtual std::shared_ptr<RegularizePtdwAgent>
-  CreateRegularizePtdwAgent(const Batch& batch, const ProcessBatchesArgs& args, double tau);
+  CreateRegularizePtdwAgent(const Batch& batch, const ProcessBatchesArgs& args, float tau);
 
   virtual bool Reconfigure(const RegularizerConfig& config);
