+++ src/artm/regularizer/topic_selection_theta.h
@@ -51,7 +51,7 @@ class TopicSelectionTheta : public RegularizerInterface {
   explicit TopicSelectionTheta(const TopicSelectionThetaConfig& config);
 
   virtual std::shared_ptr<RegularizeThetaAgent>
-  CreateRegularizeThetaAgent(const Batch& batch, const ProcessBatchesArgs& args, double tau);
+  CreateRegularizeThetaAgent(const Batch& batch, const ProcessBatchesArgs& args, float tau);
 
   virtual google::protobuf::RepeatedPtrField<std::string> topics_to_regularize();
