+++ src/artm/score/topic_kernel.cc
@@ -34,7 +34,7 @@ std::shared_ptr<Score> TopicKernel::CalculateScore(const artm::core::PhiMatrix& 
   if (config_.has_class_id())
     class_id = config_.class_id();
 
-  double probability_mass_threshold = config_.probability_mass_threshold();
+  float probability_mass_threshold = config_.probability_mass_threshold();
   if (probability_mass_threshold < 0 || probability_mass_threshold > 1) {
     BOOST_THROW_EXCEPTION(artm::core::ArgumentOutOfRangeException(
         "TopicKernelScoreConfig.probablility_mass_threshold",
@@ -53,44 +53,44 @@ std::shared_ptr<Score> TopicKernel::CalculateScore(const artm::core::PhiMatrix& 
 
   for (int topic_index = 0; topic_index < topic_size; ++topic_index) {
     if (topics_to_score[topic_index]) {
-      kernel_size->Add(0.0);
-      kernel_purity->Add(0.0);
-      kernel_contrast->Add(0.0);
-      kernel_coherence->Add(0.0);
+      kernel_size->Add(0.0f);
+      kernel_purity->Add(0.0f);
+      kernel_contrast->Add(0.0f);
+      kernel_coherence->Add(0.0f);
 
       topic_kernel_score->add_topic_name(topic_name.Get(topic_index));
     } else {
-      kernel_size->Add(-1);
-      kernel_purity->Add(-1);
-      kernel_contrast->Add(-1);
-      kernel_coherence->Add(-1);
+      kernel_size->Add(-1.0f);
+      kernel_purity->Add(-1.0f);
+      kernel_contrast->Add(-1.0f);
+      kernel_coherence->Add(-1.0f);
     }
   }
 
   auto n_wt = GetPhiMatrix(instance_->config()->nwt_name());
-  std::vector<double> n_t(topic_size, 0.0);
+  std::vector<float> n_t(topic_size, 0.0f);
   std::vector<std::vector<core::Token> > topic_kernel_tokens(
       topic_size, std::vector<core::Token>());
 
   for (int topic_index = 0; topic_index < topic_size; ++topic_index)
     for (int token_index = 0; token_index < token_size; ++token_index)
-      n_t[topic_index] += static_cast<double>(n_wt->get(token_index, topic_index));
+      n_t[topic_index] += n_wt->get(token_index, topic_index);
 
   for (int token_index = 0; token_index < token_size; ++token_index) {
     if (p_wt.token(token_index).class_id == class_id) {
-      double p_w = 0.0;
+      float p_w = 0.0;
       for (int topic_index = 0; topic_index < topic_size; ++topic_index) {
         if (topics_to_score[topic_index])
-          p_w += static_cast<double>(p_wt.get(token_index, topic_index) * n_t[topic_index]);
+          p_w += p_wt.get(token_index, topic_index) * n_t[topic_index];
       }
 
       for (int topic_index = 0; topic_index < topic_size; ++topic_index) {
         if (topics_to_score[topic_index]) {
-          double value = static_cast<double>(p_wt.get(token_index, topic_index));
-          double p_tw = (p_w > 0.0) ? (value * n_t[topic_index] / p_w) : 0.0;
+          float value = p_wt.get(token_index, topic_index);
+          float p_tw = (p_w > 0.0f) ? (value * n_t[topic_index] / p_w) : 0.0f;
 
           if (p_tw >= probability_mass_threshold) {
-            artm::core::repeated_field_append(kernel_size, topic_index, 1.0);
+            artm::core::repeated_field_append(kernel_size, topic_index, 1.0f);
             artm::core::repeated_field_append(kernel_purity, topic_index, value);
             artm::core::repeated_field_append(kernel_contrast, topic_index, p_tw);
             topic_kernel_tokens[topic_index].push_back(p_wt.token(token_index));
@@ -102,7 +102,7 @@ std::shared_ptr<Score> TopicKernel::CalculateScore(const artm::core::PhiMatrix& 
 
   // contrast = sum(p(t|w)) / kernel_size
   for (int topic_index = 0; topic_index < topic_size; ++topic_index) {
-    double value = 0;
+    float value = 0.0f;
     if (kernel_size->Get(topic_index) > config_.eps()) {
       value = kernel_contrast->Get(topic_index) / kernel_size->Get(topic_index);
     }
@@ -122,14 +122,14 @@ std::shared_ptr<Score> TopicKernel::CalculateScore(const artm::core::PhiMatrix& 
     average_kernel_coherence /= average_kernel_coherence > 0 ? denominator : 1;
   }
 
-  double average_kernel_size = 0.0;
-  double average_kernel_purity = 0.0;
-  double average_kernel_contrast = 0.0;
-  double useful_topics_count = 0.0;
+  float average_kernel_size = 0.0f;
+  float average_kernel_purity = 0.0f;
+  float average_kernel_contrast = 0.0f;
+  float useful_topics_count = 0.0f;
   auto kernel_tokens = topic_kernel_score->mutable_kernel_tokens();
 
   for (int topic_index = 0; topic_index < topic_size; ++topic_index) {
-    double current_kernel_size = kernel_size->Get(topic_index);
+    float current_kernel_size = kernel_size->Get(topic_index);
     bool useful_topic = (current_kernel_size != -1);
     if (useful_topic) {
       useful_topics_count += 1;
